Use shared resolveBackAction() in FrontendScreen
Move BackAction/resolveBackAction into WebViewBackNavigation.kt
(shared with WebViewActivity) and remove duplicated logic from
FrontendScreen. Tests moved to WebViewBackNavigationTest.
